move types to its own file - update path to cli binary in wrapper
